Quantity and calories now use double type instead of long

+++ src/main/java/net/yorksolutions/peternepomucenopantrybe/DTOs/CookedRecipeDTO.java
@@ -7,6 +7,6 @@ public class CookedRecipeDTO {
     public String name;
     public String image;
     public String description;
-    public Long calories;
+    public Double calories;
     public Double weight;
 }
